HACK: Honour cursor-size config
By scaling the cursor surface.

+++ libweston/libweston-internal.h
@@ -583,7 +583,7 @@ void
 convert_size_by_transform_scale(int32_t *width_out, int32_t *height_out,
 				int32_t width, int32_t height,
 				uint32_t transform,
-				int32_t scale);
+				float scale);
 
 /* User authentication for remote backends */
